Better error management and improvemet to tasks schduler

+++ get_episodes.py
@@ -21,6 +21,7 @@ def update_all_episodes():
         base_url_sonarr = ""
     else:
         base_url_sonarr = "/" + config_sonarr[2].strip("/")
+    apikey_sonarr = config_sonarr[4]
 
     # Get current episodes id in DB
     current_episodes_db = c.execute('SELECT sonarrEpisodeId FROM table_episodes').fetchall()
@@ -32,7 +33,7 @@ def update_all_episodes():
     seriesIdList = c.fetchall()
     for seriesId in seriesIdList:
         # Get episodes data for a series from Sonarr
-        url_sonarr_api_episode = protocol_sonarr + "://" + config_sonarr[0] + ":" + str(config_sonarr[1]) + base_url_sonarr + "/api/episode?seriesId=" + str(seriesId[0]) + "&apikey=" + config_sonarr[4]
+        url_sonarr_api_episode = protocol_sonarr + "://" + config_sonarr[0] + ":" + str(config_sonarr[1]) + base_url_sonarr + "/api/episode?seriesId=" + str(seriesId[0]) + "&apikey=" + apikey_sonarr
         r = requests.get(url_sonarr_api_episode)
         for episode in r.json():
             if episode['hasFile']:
@@ -80,47 +81,52 @@ def add_new_episodes():
         base_url_sonarr = ""
     else:
         base_url_sonarr = "/" + config_sonarr[2].strip("/")
+    apikey_sonarr = config_sonarr[4]
 
-    # Get current episodes in DB
-    current_episodes_db = c.execute('SELECT sonarrEpisodeId FROM table_episodes').fetchall()
-    current_episodes_db_list = [x[0] for x in current_episodes_db]
-    current_episodes_sonarr = []
+    if apikey_sonarr == None:
+        # Close database connection
+        c.close()
+        pass
+    else:
+        # Get current episodes in DB
+        current_episodes_db = c.execute('SELECT sonarrEpisodeId FROM table_episodes').fetchall()
+        current_episodes_db_list = [x[0] for x in current_episodes_db]
+        current_episodes_sonarr = []
 
-    # Get sonarrId for each series from database
-    c.execute("SELECT sonarrSeriesId FROM table_shows")
-    seriesIdList = c.fetchall()
-    for seriesId in seriesIdList:
-        # Get episodes data for a series from Sonarr
-        url_sonarr_api_episode = protocol_sonarr + "://" + config_sonarr[0] + ":" + str(config_sonarr[1]) + base_url_sonarr + "/api/episode?seriesId=" + str(seriesId[0]) + "&apikey=" + config_sonarr[4]
-        r = requests.get(url_sonarr_api_episode)
+        # Get sonarrId for each series from database
+        c.execute("SELECT sonarrSeriesId FROM table_shows")
+        seriesIdList = c.fetchall()
+        for seriesId in seriesIdList:
+            # Get episodes data for a series from Sonarr
+            url_sonarr_api_episode = protocol_sonarr + "://" + config_sonarr[0] + ":" + str(config_sonarr[1]) + base_url_sonarr + "/api/episode?seriesId=" + str(seriesId[0]) + "&apikey=" + apikey_sonarr
+            r = requests.get(url_sonarr_api_episode)
+            for episode in r.json():
+                if episode['hasFile']:
+                    # Add shows in Sonarr to current shows list
+                    current_episodes_sonarr.append(episode['id'])
 
-        for episode in r.json():
-            if episode['hasFile']:
-                # Add shows in Sonarr to current shows list
-                current_episodes_sonarr.append(episode['id'])
+                    try:
+                        c.execute('''INSERT INTO table_episodes(sonarrSeriesId, sonarrEpisodeId, title, path, season, episode) VALUES (?, ?, ?, ?, ?, ?)''', (episode['seriesId'], episode['id'], episode['title'], episode['episodeFile']['path'], episode['seasonNumber'], episode['episodeNumber']))
+                    except:
+                        pass
+            db.commit()
 
-                try:
-                    c.execute('''INSERT INTO table_episodes(sonarrSeriesId, sonarrEpisodeId, title, path, season, episode) VALUES (?, ?, ?, ?, ?, ?)''', (episode['seriesId'], episode['id'], episode['title'], episode['episodeFile']['path'], episode['seasonNumber'], episode['episodeNumber']))
-                except:
-                    pass
+        # Delete episodes not in Sonarr anymore
+        deleted_items = []
+        for item in current_episodes_db_list:
+            if item not in current_episodes_sonarr:
+                deleted_items.append(tuple([item]))
+        c.executemany('DELETE FROM table_episodes WHERE sonarrEpisodeId = ?',deleted_items)
+        
+        # Commit changes to database table
         db.commit()
 
-    # Delete episodes not in Sonarr anymore
-    deleted_items = []
-    for item in current_episodes_db_list:
-        if item not in current_episodes_sonarr:
-            deleted_items.append(tuple([item]))
-    c.executemany('DELETE FROM table_episodes WHERE sonarrEpisodeId = ?',deleted_items)
-    
-    # Commit changes to database table
-    db.commit()
+        # Close database connection
+        c.close()
 
-    # Close database connection
-    c.close()
-
-    # Store substitles from episodes we've just added
-    new_scan_subtitles()
-    try:
-        list_missing_subtitles()
-    except:
-        pass
+        # Store substitles from episodes we've just added
+        new_scan_subtitles()
+        try:
+            list_missing_subtitles()
+        except:
+            pass
